+++ src/__internal__/utils/helpers/guid/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Generates a random guid, useful for creating unique IDs.
+ */
+export default (): string => {
+  function s4(): string {
+    return Math.floor((1 + Math.random()) * 0x10000)
+      .toString(16)
+      .substring(1);
+  }
+  return `${s4()}${s4()}-${s4()}-${s4()}-${s4()}-${s4()}${s4()}${s4()}`;
+};