+++ src/__spec_helper__/mock-match-media.d.ts
@@ -1,4 +1,6 @@
 declare function setup(): void;
-declare function mockMatchMedia(matches?: boolean): jest.Mock;
+declare function mockMatchMedia(
+  matches?: boolean
+): { removeListener: jest.Mock };
 
 export { setup, mockMatchMedia };