+++ src/components/button/button.config.js
@@ -1,9 +1,0 @@
-export const BUTTON_VARIANTS = [
-  "primary",
-  "secondary",
-  "tertiary",
-  "dashed",
-  "darkBackground",
-];
-export const BUTTON_SIZES = ["small", "medium", "large"];
-export const BUTTON_ICON_POSITIONS = ["before", "after"];