+++ src/components/button/button.config.ts
@@ -1,0 +1,11 @@
+import { ButtonTypes } from "./button.component";
+
+export const BUTTON_VARIANTS: ButtonTypes[] = [
+  "primary",
+  "secondary",
+  "tertiary",
+  "dashed",
+  "darkBackground",
+];
+export const BUTTON_SIZES = ["small", "medium", "large"];
+export const BUTTON_ICON_POSITIONS = ["before", "after"];