+++ src/components/button/index.ts
@@ -1,0 +1,2 @@
+export { default, ButtonWithForwardRef } from "./button.component";
+export type { ButtonProps } from "./button.component";