+++ src/components/icon/index.d.ts
@@ -1,1 +1,2 @@
 export { default } from "./icon";
+export * from "./icon";