+++ src/locales/index.ts
@@ -1,0 +1,3 @@
+export type { default } from "./locale";
+export { default as EnGb } from "./en-gb";
+export { default as PlPl } from "./pl-pl";